memory optimization (#601)
* update ethash to cache seeds

* add pprof

* update cache function to us epoch as param instead of block hight

* fix bug

* remove rootHeightToHashes

* optimize the heightToMinorBlockHashes map

* update getBlockCountByHeight return result for not cache height

* add root block to cache when write root block to db

* add gc api

* rocksdb->leveldb

* add SetMining

* fix miner

* first commit
1.modify ethash
2.modify minorblockchain.go map
3.fix miner goroutinue

* update heightToMBlockHashCount

* rocksdb->leveldb

* 1.delete fmt
2.delete pprof

* delete empty line

* modify beyond pr

* modify leveldb

* update HandleNewTip()

* delete miner's interface.go IsRemoteMining

* update README(#602)

+++ core/shardstate_test.go
@@ -814,10 +814,10 @@ func TestStaleBlockCount(t *testing.T) {
 
 	_, _, err = shardState.FinalizeAndAddBlock(b1)
 	checkErr(err)
-	assert.Equal(t, shardState.getBlockCountByHeight(1), uint64(1))
+	assert.Equal(t, shardState.getBlockCountByHeight(1), 1)
 	_, _, err = shardState.FinalizeAndAddBlock(b22)
 	checkErr(err)
-	assert.Equal(t, shardState.getBlockCountByHeight(1), uint64(2))
+	assert.Equal(t, shardState.getBlockCountByHeight(1), 2)
 }
 
 func TestXShardTxSent(t *testing.T) {
